Move to ldap auth mapper
Fixes some ts and lint issues, also revert changes to tsconfig
